requirements.txt
fastapi
uvicorn
pymongo
redis
crewai
python-dotenv
pydantic[email]
google-generativeai
litellm
gunicorn